+++ slack-market/commands/quote.rb
@@ -1,7 +1,8 @@
 module SlackMarket
   module Commands
     class Quote < SlackRubyBot::Commands::Base
-      scan(/(\b[A-Z]{2,}+\b)/) do |client, data, stocks|
+      scan(/\b[A-Z]{2,}\b|\$[A-Z]{1,}\b|\b[A-Z]{1,}\$/) do |client, data, stocks|
+        stocks = stocks.flatten.map { |s| s.tr('$', '') }
         YahooFinance::Client.new.quotes(stocks, [:name, :symbol, :last_trade_price, :change, :change_in_percent]).each do |quote|
           next if quote.name == 'N/A'
           logger.info "#{client.team}, user=#{data.user} - #{quote.name} (#{quote.symbol}): $#{quote.last_trade_price}"