+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Changelog
 
+* 2/5/2016: Respond to single character stocks with a $ sign, eg. `$F` - [@dblock](https://github.com/dblock).
 * 2/5/2016: Don't respond to channel mentions that contain what looks like a ticker - [@dblock](https://github.com/dblock).
 * 2/4/2016: Initial public release - [@dblock](https://github.com/dblock).